fix: 레이아웃 관련 빌드에러 수정
