+++ src/app/partner/mypage/preview/@tabs/layout.tsx
@@ -1,5 +1,18 @@
 import React from 'react'
 
+import PartnerProfileHeader from '@/components/pages/partner/profile/PartnerProfileHeader'
+import PartnerProfilePreviewTabs from '@/components/pages/partner/mypage/preview/PartnerProfilePreviewTabs'
+import PartnerProfileTopInfo from '@/components/pages/partner/profile/PartnerProfileTopInfo'
+
 export default function layout({ children }: { children: React.ReactNode }) {
-  return <div className="p-7 min-h-[90vh] bg-white w-full">{children}</div>
+  return (
+    <>
+      <PartnerProfileHeader />
+      <main className="mt-[50px] w-full relative z-[20]">
+        <PartnerProfileTopInfo />
+        <PartnerProfilePreviewTabs />
+        <div className="p-7 min-h-[90vh] bg-white w-full">{children}</div>
+      </main>
+    </>
+  )
 }
